Improve `yarn dev` DX in `block-scripts`
