+++ packages/block-scripts/scripts/build.js
@@ -1,41 +1,19 @@
 import chalk from "chalk";
 import { promisify } from "node:util";
 import webpack from "webpack";
-import { BundleAnalyzerPlugin } from "webpack-bundle-analyzer";
 
 import { cleanDist } from "../shared/clean-dist.js";
-import { extractBlockScriptsConfigFromPackageJson } from "../shared/config.js";
-import { generateBaseWebpackConfig } from "../shared/generate-base-webpack-config.js";
 import { blockDistDirPath } from "../shared/paths.js";
+import { generateBuildWebpackConfig } from "../shared/webpack-config.js";
 
 const promisifiedWebpack = promisify(webpack);
 
 const script = async () => {
   await cleanDist();
 
-  const baseWebpackConfig = await generateBaseWebpackConfig("production");
-
-  const analyze = Boolean(
-    ["true", "1"].includes(process.env.ANALYZE) ||
-      (await extractBlockScriptsConfigFromPackageJson()).analyze,
-  );
-
-  /** @type import("webpack").Configuration */
-  const webpackConfig = {
-    ...baseWebpackConfig,
-    plugins: analyze
-      ? [
-          ...baseWebpackConfig.plugins,
-          new BundleAnalyzerPlugin({
-            analyzerMode: "static",
-            openAnalyzer: false,
-            reportFilename: "webpack-bundle-analyzer-report.html",
-          }),
-        ]
-      : baseWebpackConfig.plugins,
-  };
-
-  const stats = await promisifiedWebpack([webpackConfig]);
+  const stats = await promisifiedWebpack([
+    await generateBuildWebpackConfig("production"),
+  ]);
 
   if (stats.hasErrors()) {
     console.log(stats.toString());
